Refactor: 팔로우팔로잉 페이지 헤더 수정 및 불필요 파일 삭제
