--- src/app/user/[userId]/(follow)/_components/Header.tsx
@@ -1,20 +1,0 @@
-'use client';
-import { useRouter, useParams } from 'next/navigation';
-import CommonHeader from '@/components/Header/Header';
-
-interface HeaderProps {
-  title: string;
-}
-
-function Header({ title }: HeaderProps) {
-  const router = useRouter();
-  const param = useParams<{ userId: string }>();
-
-  const handleBackButtonClick = () => {
-    router.push(`/user/${param?.userId}/mylist`);
-  };
-
-  return <CommonHeader left={'back'} leftClick={handleBackButtonClick} title={title} />;
-}
-
-export default Header;
