--- src/app/user/[userId]/(follow)/_components/mockData.ts
@@ -1,203 +1,0 @@
-interface UserProfileType {
-  id: number;
-  profileImageUrl: string;
-  nickname: string;
-}
-
-export const mockFollowers: UserProfileType[] = [
-  {
-    id: 1,
-    profileImageUrl: 'https://randomuser.me/api/portraits/men/1.jpg',
-    nickname: 'JohnDoe1',
-  },
-  {
-    id: 2,
-    profileImageUrl: 'https://randomuser.me/api/portraits/women/2.jpg',
-    nickname: 'JaneDoe2',
-  },
-  {
-    id: 3,
-    profileImageUrl: 'https://randomuser.me/api/portraits/men/3.jpg',
-    nickname: 'BobSmith3',
-  },
-  {
-    id: 4,
-    profileImageUrl: 'https://randomuser.me/api/portraits/women/4.jpg',
-    nickname: 'AliceJohnson4',
-  },
-  {
-    id: 5,
-    profileImageUrl: 'https://randomuser.me/api/portraits/men/5.jpg',
-    nickname: 'CharlieBrown5',
-  },
-  {
-    id: 6,
-    profileImageUrl: 'https://randomuser.me/api/portraits/women/6.jpg',
-    nickname: 'EmmaWatson6',
-  },
-  {
-    id: 7,
-    profileImageUrl: 'https://randomuser.me/api/portraits/men/7.jpg',
-    nickname: 'DavidMiller7',
-  },
-  {
-    id: 8,
-    profileImageUrl: 'https://randomuser.me/api/portraits/women/8.jpg',
-    nickname: 'OliviaSmith8',
-  },
-  {
-    id: 9,
-    profileImageUrl: 'https://randomuser.me/api/portraits/men/9.jpg',
-    nickname: 'SophiaJohnson9',
-  },
-  {
-    id: 10,
-    profileImageUrl: 'https://randomuser.me/api/portraits/women/10.jpg',
-    nickname: 'LiamBrown10',
-  },
-  {
-    id: 11,
-    profileImageUrl: 'https://randomuser.me/api/portraits/men/11.jpg',
-    nickname: 'AvaDavis11',
-  },
-  {
-    id: 12,
-    profileImageUrl: 'https://randomuser.me/api/portraits/women/12.jpg',
-    nickname: 'NoahWilliams12',
-  },
-  {
-    id: 13,
-    profileImageUrl: 'https://randomuser.me/api/portraits/men/13.jpg',
-    nickname: 'IsabellaMoore13',
-  },
-  {
-    id: 14,
-    profileImageUrl: 'https://randomuser.me/api/portraits/women/14.jpg',
-    nickname: 'LucasTaylor14',
-  },
-  {
-    id: 15,
-    profileImageUrl: 'https://randomuser.me/api/portraits/men/15.jpg',
-    nickname: 'MiaAnderson15',
-  },
-  {
-    id: 16,
-    profileImageUrl: 'https://randomuser.me/api/portraits/women/16.jpg',
-    nickname: 'EthanThomas16',
-  },
-  {
-    id: 17,
-    profileImageUrl: 'https://randomuser.me/api/portraits/men/17.jpg',
-    nickname: 'AriaMartinez17',
-  },
-  {
-    id: 18,
-    profileImageUrl: 'https://randomuser.me/api/portraits/women/18.jpg',
-    nickname: 'LoganJackson18',
-  },
-  {
-    id: 19,
-    profileImageUrl: 'https://randomuser.me/api/portraits/men/19.jpg',
-    nickname: 'ChloeWhite19',
-  },
-  {
-    id: 20,
-    profileImageUrl: 'https://randomuser.me/api/portraits/women/20.jpg',
-    nickname: 'ElijahHill20',
-  },
-  {
-    id: 21,
-    profileImageUrl: 'https://randomuser.me/api/portraits/men/21.jpg',
-    nickname: 'AbigailCarter21',
-  },
-  {
-    id: 22,
-    profileImageUrl: 'https://randomuser.me/api/portraits/women/22.jpg',
-    nickname: 'CarterLee22',
-  },
-  {
-    id: 23,
-    profileImageUrl: 'https://randomuser.me/api/portraits/men/23.jpg',
-    nickname: 'GraceKing23',
-  },
-  {
-    id: 24,
-    profileImageUrl: 'https://randomuser.me/api/portraits/women/24.jpg',
-    nickname: 'JacksonScott24',
-  },
-  {
-    id: 25,
-    profileImageUrl: 'https://randomuser.me/api/portraits/men/25.jpg',
-    nickname: 'LilyAdams25',
-  },
-  {
-    id: 26,
-    profileImageUrl: 'https://randomuser.me/api/portraits/women/26.jpg',
-    nickname: 'AidenWilson26',
-  },
-  {
-    id: 27,
-    profileImageUrl: 'https://randomuser.me/api/portraits/men/27.jpg',
-    nickname: 'ZoeEvans27',
-  },
-  {
-    id: 28,
-    profileImageUrl: 'https://randomuser.me/api/portraits/women/28.jpg',
-    nickname: 'MasonHall28',
-  },
-  {
-    id: 29,
-    profileImageUrl: 'https://randomuser.me/api/portraits/men/29.jpg',
-    nickname: 'MadisonWard29',
-  },
-  {
-    id: 30,
-    profileImageUrl: 'https://randomuser.me/api/portraits/women/30.jpg',
-    nickname: 'EzraBrooks30',
-  },
-  {
-    id: 31,
-    profileImageUrl: 'https://randomuser.me/api/portraits/men/31.jpg',
-    nickname: 'AveryBennett31',
-  },
-  {
-    id: 32,
-    profileImageUrl: 'https://randomuser.me/api/portraits/women/32.jpg',
-    nickname: 'WyattMiller32',
-  },
-  {
-    id: 33,
-    profileImageUrl: 'https://randomuser.me/api/portraits/men/33.jpg',
-    nickname: 'ScarlettFisher33',
-  },
-  {
-    id: 34,
-    profileImageUrl: 'https://randomuser.me/api/portraits/women/34.jpg',
-    nickname: 'NathanSimmons34',
-  },
-  {
-    id: 35,
-    profileImageUrl: 'https://randomuser.me/api/portraits/men/35.jpg',
-    nickname: 'HannahLopez35',
-  },
-  {
-    id: 36,
-    profileImageUrl: 'https://randomuser.me/api/portraits/women/36.jpg',
-    nickname: 'JackCooper36',
-  },
-  {
-    id: 37,
-    profileImageUrl: 'https://randomuser.me/api/portraits/men/37.jpg',
-    nickname: 'AddisonFoster37',
-  },
-  {
-    id: 38,
-    profileImageUrl: 'https://randomuser.me/api/portraits/women/38.jpg',
-    nickname: 'AndrewMurray38',
-  },
-  {
-    id: 39,
-    profileImageUrl: 'https://randomuser.me/api/portraits/men/39.jpg',
-    nickname: 'Joe',
-  },
-];
